+++ database/seed.js
@@ -1,0 +1,56 @@
+const db  = require('./index.js');
+const Question = require('./Question.js');
+
+const sampleData = [
+    {
+question_id: 058,
+product_id: 017,
+user: 'BobbyB',
+question_body: 'Does this product fit both men and women?',
+answer: [
+    {
+    id: 023,
+    body: 'Yes it does.',
+    user: 'REI Helper Sally',
+    helpful: {
+        yes: 0,
+        no: 1,
+    },
+    },
+],
+}, 
+{
+    question_id: 009,
+    product_id: 043,
+    user: 'Rachel098',
+    question_body: 'How durable is this?',
+    answer: [
+        {
+        id: 023,
+        body: 'It can stand up to all your favortie activities.',
+        user: 'REI Helper Jim',
+        helpful: {
+            yes: 2,
+            no: 0,
+        },
+        }, 
+        {
+            id: 079,
+            body: 'It tore after one week.',
+            user: 'CaliBro43',
+            helpful: {
+                yes: 1,
+                no: 3,
+            },
+            },
+    ],
+    },
+]
+
+const insertSampleData = () => {
+Question.create(sampleData)
+    .then(() => db.close())
+    .catch(err => console.log(err));
+};
+
+insertSampleData();
