+++ package.json
@@ -11,7 +11,7 @@
   "scripts": {
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
-    "db-seed": "node database/index.js"
+    "db-seed": "node database/seed.js"
   },
   "dependencies": {
     "axios": "^0.20.0",
